+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+### 0.1.0+268b900 (Released 2022-3-31)
+* Additions:
+    * latest commit #268b900
+    * [[#268b900](https://github.com/CSBiology/FsSpreadsheet/commit/268b90012e90a038ba9bfbad6ef4ff6d8f3ebda4)] add columnbuilder tp dsl
+    * [[#eed6806](https://github.com/CSBiology/FsSpreadsheet/commit/eed680645d473e51a3bb47b9d9db20d0e8eb5f6b)] add sheet and workbook builder to DSL
+    * [[#ed45200](https://github.com/CSBiology/FsSpreadsheet/commit/ed452001d696457e881d9a5bd390fb58455fa3de)] add optional and required cells to DSL
+
 ### 0.0.2+2f09131 (Released 2022-1-21)
 * Additions:
     * latest commit #2f09131